test: Fixture 분리.
